fixed click handler for video gallery items
This commit fixes the `closeOnSlideClick` condition of the `handleClick`
function for video gallery items. `closeOnSlideClick` is only reached if the
`parent` is equal to the slide container (`.slides`).

The `parent` is set to the parentNode of
the event target, which is fine for image gallery items since they
adhere to the following general DOM structure:

```
<div class="slides">
  <div class="slide">
    <img src="...">
  </div>
</div>
```

However, this is problematic for video gallery items since they adhere
to the following general DOM structure:

```
<div class="slides">
  <div class="slide">
    <div class="video-content slide-content">
      <img src="...">
    </div>
  </div>
</div>
```

In that scenario, the click event can be triggered on the `.video-content`
div rather than the `.slide` div. This commit looks for the closest
`.slides` container of the click event target and sets that to parent.