Allow data attributes to use any name
Some DOM element properties (e.g. the title or id properties) are defined on the HTMLElement prototype as a DOMString which prevents them from ever evaluating as undefined.
This causes any dataset attributes with names overlapping these inherited properties to be ignored by getItemProperty().
When determining the value of an item property we must first check for the existence of an explicitly defined dataset attribute so that any name can be used.